Swap indexes in trait to allow project specific indexes to be swapped too
