+++ src/Commands/IndexProductsCommand.php
@@ -11,9 +11,12 @@ use Rapidez\Core\Facades\Rapidez;
 use Rapidez\Core\Jobs\IndexProductJob;
 use Rapidez\Core\Models\Category;
 use TorMorten\Eventy\Facades\Eventy;
+use Rapidez\Core\Commands\Traits\SwapIndexes;
 
 class IndexProductsCommand extends Command
 {
+    use SwapIndexes;
+
     protected $signature = 'rapidez:index {store? : Store ID from Magento}';
 
     protected $description = 'Index the products in Elasticsearch';
@@ -43,7 +46,19 @@ class IndexProductsCommand extends Command
 
             $alias = config('rapidez.es_prefix').'_products_'.$store->store_id;
             $index = $alias.'_'.Carbon::now()->format('YmdHis');
-            $this->createIndex($index);
+            $this->createIndex($index, Eventy::filter('index.product.mapping', [
+                'properties' => [
+                    'price' => [
+                        'type' => 'double',
+                    ],
+                    'children' => [
+                        'type' => 'flattened',
+                    ],
+                    'grouped' => [
+                        'type' => 'flattened',
+                    ],
+                ],
+            ]));
 
             try {
                 $flat = (new $productModel())->getTable();
@@ -101,46 +116,5 @@ class IndexProductsCommand extends Command
         }
 
         return $data;
-    }
-
-    public function switchAlias(string $alias, string $index): void
-    {
-        $this->elasticsearch->indices()->putAlias([
-            'index' => $index,
-            'name'  => $alias,
-        ]);
-
-        $aliases = $this->elasticsearch->indices()->getAlias(['name' => $alias]);
-        foreach ($aliases as $indexLinkedToAlias => $aliasData) {
-            if ($indexLinkedToAlias != $index) {
-                $this->elasticsearch->indices()->deleteAlias([
-                    'index' => $indexLinkedToAlias,
-                    'name'  => $alias,
-                ]);
-                $this->elasticsearch->indices()->delete(['index' => $indexLinkedToAlias]);
-            }
-        }
-    }
-
-    public function createIndex(string $index): void
-    {
-        $this->elasticsearch->indices()->create([
-            'index' => $index,
-            'body'  => [
-                'mappings' => Eventy::filter('index.product.mapping', [
-                    'properties' => [
-                        'price' => [
-                            'type' => 'double',
-                        ],
-                        'children' => [
-                            'type' => 'flattened',
-                        ],
-                        'grouped' => [
-                            'type' => 'flattened',
-                        ],
-                    ],
-                ]),
-            ],
-        ]);
     }
 }
